+++ CHANGELOG.md
@@ -6,6 +6,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## Documentation
+- Document `People` page property (#162)
+
 ## [1.3.0]
 
 ## Added