+++ CHANGELOG.md
@@ -6,8 +6,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## Added
+- Support to comments (#163)
+
 ## Fixed
 - Update options without color (#166)
+
 ## Documentation
 - Document `People` page property (#162)