+++ CHANGELOG.md
@@ -6,6 +6,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [1.4.0]
+
 ## Added
 - Support to comments (#163)
 
@@ -262,4 +264,5 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 [1.1.0]: https://github.com/mariosimao/notion-sdk-php/releases/tag/v1.1.0
 [1.2.0]: https://github.com/mariosimao/notion-sdk-php/releases/tag/v1.2.0
 [1.3.0]: https://github.com/mariosimao/notion-sdk-php/releases/tag/v1.3.0
-[Unreleased]: https://github.com/mariosimao/notion-sdk-php/compare/v1.3.0...HEAD
+[1.3.0]: https://github.com/mariosimao/notion-sdk-php/releases/tag/v1.4.0
+[Unreleased]: https://github.com/mariosimao/notion-sdk-php/compare/v1.4.0...HEAD